funding goal changed to a positive field

+++ crowd_surfer/models.py
@@ -14,12 +14,12 @@ class Project(models.Model):
     owner = models.ForeignKey(User, on_delete=models.CASCADE)
     funding_start_date = models.DateField() 
     funding_end_date = models.DateField()
-    funding_goal = models.IntegerField() 
+    funding_goal = models.PositiveIntegerField()
 
 class Reward(models.Model): 
     reward_type = models.CharField(max_length=255)
     reward_item = models.CharField(max_length=255)
-    reward_amount = models.FloatField() 
+    reward_amount = models.PositiveIntegerField() 
 
 class Comment(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE) 
@@ -28,7 +28,7 @@ class Comment(models.Model):
     text = models.TextField() 
 
 class Donation(models.Model): 
-    amount = models.FloatField() 
+    amount = models.PositiveIntegerField() 
     reward = models.ForeignKey(Reward, on_delete=models.CASCADE)
     user = models.ForeignKey(User, on_delete=models.CASCADE)
